api-model-pneumonia/requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu==2.*
pillow
numpy
python-multipart
